# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Splits API
  version: 0.1.0
  description: |
    REST API for Splits, an expense-sharing app. Health checks live at the
    root; every other endpoint is under `/v1`.

    Every response carries an `X-Request-ID` header: the request's own
    `X-Request-ID` when it sends a valid one (1–64 letters, digits, `-` or
    `_`), otherwise one the server generates. Server logs carry the same ID.
    Request bodies are limited to 64 KB (`request-too-large`).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (Docker Compose)
...
security: []
paths:
...
  /v1/groups/{groupId}/expenses/preview:
    parameters:
      - $ref: "#/components/parameters/GroupId"
    post:
      operationId: previewExpense
      summary: Preview an Expense's Shares
      description: |
        Computes the exact Shares an Expense would get, without saving it
        (FR-E4, ADR-0006). The same code computes them on save, so the
        preview always matches. Reads only, so it takes no Idempotency-Key.
        Refuses the input exactly as creating would.
      x-splits-read-only: true
      tags: [expenses]
      security:
        - bearerAuth: []
      requestBody:
        description: The Expense as it would be saved.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
            example:
              payer_member_id: 0190b6c4-0000-7000-8000-0000000000d1
              amount: {minor: 100001, currency: INR}
              category: food_drink
              note: Dinner at the beach shack
              spent_on: "2026-10-01"
              split:
                method: equal
                members:
                  - member_id: 0190b6c4-0000-7000-8000-0000000000d1
                  - member_id: 0190b6c4-0000-7000-8000-0000000000d2
                  - member_id: 0190b6c4-0000-7000-8000-0000000000d3
      responses:
        "200":
          description: The Shares the Expense would get.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpensePreview"
              example:
                amount: {minor: 100001, currency: INR}
                shares:
                  - member_id: 0190b6c4-0000-7000-8000-0000000000d1
                    share: {minor: 33334, currency: INR}
                  - member_id: 0190b6c4-0000-7000-8000-0000000000d2
                    share: {minor: 33334, currency: INR}
                  - member_id: 0190b6c4-0000-7000-8000-0000000000d3
                    share: {minor: 33333, currency: INR}
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        An opaque access token from an AuthSession (ADR-0011): 32 random
        bytes, base64url, valid 15 minutes. Until the User's email is
        verified it works for `GET /v1/me` only; other endpoints answer
        `email-not-verified`.
  parameters:
    GroupId:
      name: groupId
      in: path
      required: true
      description: The Group's ID.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Problem:
      description: |
        RFC 9457 problem details. `type` is a stable URI,
        `https://splits.dev/problems/<slug>`, that clients map to a localized
        message. Clients must treat an unknown `type` as a generic error.
        Slugs in use:
        - `not-ready`: the server cannot reach its database (503).
        - `invalid-request`: the request could not be decoded (400).
        - `validation-failed`: one or more fields are invalid; see `errors` (400).
        - `invalid-code`: the one-time code is wrong, expired, used up or unknown (400).
        - `idempotency-key-required`: a write by a signed-in User has no valid `Idempotency-Key` (400).
        - `unauthenticated`: no access token, or it is unknown, expired or revoked (401).
        - `invalid-credentials`: the email or password is wrong (401).
        - `email-not-verified`: the User must verify their email first (403).
        - `email-taken`: a verified User already has this email (409).
        - `idempotency-key-in-progress`: a request with the same `Idempotency-Key` is still being processed (409).
        - `request-too-large`: the request body is over 64 KB (413).
        - `idempotency-key-reused`: the `Idempotency-Key` was used for a different request (422).
        - `too-many-attempts`: too many failed sign-ins; retry after `Retry-After` seconds (429).
        - `admin-required`: only an Admin of the Group may do this (403).
        - `not-found`: no such resource, or one the User can't see (404).
        - `version-conflict`: the resource changed since the `version` sent; reload it (409).
        - `group-limit-reached`: the User is already in 200 Groups (409).
        - `invalid-cursor`: the `cursor` is not one the server gave (400).
        - `member-limit-reached`: the Group already has 50 Members (409).
        - `member-not-eligible`: a Placeholder or Former Member can't be an Admin (409).
        - `group-closed`: the Group is Closed and read-only (409).

        Field error codes of an Expense (`errors[].code`): `required`,
        `invalid`, `too_long`, `not_positive`, `not_group_currency`,
        `not_a_member` (payer or Split Member isn't an active Member),
        `duplicate_member`, `no_members`, plus `ledger`'s Split reasons.
        - `internal`: an unexpected server failure (500).
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          pattern: "^https://splits\\.dev/problems/[a-z0-9-]+$"
          examples: ["https://splits.dev/problems/not-ready"]
        title:
          type: string
          description: Short, human-readable summary; not localized, not for display.
          examples: ["Service not ready"]
        status:
          type: integer
          minimum: 400
          maximum: 599
          examples: [503]
        detail:
          type: string
          description: Explanation of this occurrence; not for display.
          examples: ["The database is unreachable."]
        errors:
          description: Field validation failures, one per invalid field.
          type: array
          examples: [[{field: /amount/minor, code: required}]]
          items:
            $ref: "#/components/schemas/FieldError"
    FieldError:
      description: One invalid field in a request body.
      type: object
      required: [field, code]
      additionalProperties: false
      properties:
        field:
          type: string
          description: JSON Pointer (RFC 6901) to the invalid field in the request body.
          examples: ["/amount/minor"]
        code:
          type: string
          description: Stable machine-readable reason, such as `required` or `too_long`.
...
    Category:
      description: The fixed Category list (D7); the app localizes the names.
      type: string
      enum: [food_drink, groceries, transport, accommodation, rent, utilities, entertainment, shopping, health, travel, other]
      examples: ["food_drink"]
    SplitMethod:
      description: |
        How the Expense is divided (FR-E2). Equal only for now; exact,
        percentage and ratio follow (#18).
      type: string
      enum: [equal]
      examples: ["equal"]
...
    SplitInput:
      description: Who shares the Expense, and how.
      type: object
      required: [method, members]
      additionalProperties: false
      properties:
        method:
          $ref: "#/components/schemas/SplitMethod"
        members:
          description: 1–50 distinct active Members.
          type: array
          maxItems: 50
          items:
            type: object
            required: [member_id]
            additionalProperties: false
            properties:
              member_id:
                type: string
                format: uuid
          examples:
            - - member_id: 0190b6c4-0000-7000-8000-0000000000d1
    ExpenseInput:
      description: An Expense as entered (FR-E1). Previews and creates take the same input.
      type: object
      required: [payer_member_id, amount, category, spent_on, split]
      additionalProperties: false
      properties:
        payer_member_id:
          description: Any active Member, Placeholders included (FR-E3).
          type: string
          format: uuid
          examples: ["0190b6c4-0000-7000-8000-0000000000d1"]
        amount:
          description: "> 0, in the Group Currency for now (foreign currencies: #19)."
          $ref: "#/components/schemas/Money"
          examples:
            - {minor: 100001, currency: INR}
        category:
          $ref: "#/components/schemas/Category"
        note:
          description: Optional, at most 500 characters after trimming (D1).
          type: string
          maxLength: 2000
          examples: ["Dinner at the beach shack"]
        spent_on:
          description: The day the money was spent.
          type: string
          format: date
          examples: ["2026-10-01"]
        split:
          $ref: "#/components/schemas/SplitInput"
    ShareLine:
      description: One Member's Share.
      type: object
      required: [member_id, share]
      additionalProperties: false
      properties:
        member_id:
          type: string
          format: uuid
          examples: ["0190b6c4-0000-7000-8000-0000000000d1"]
        share:
          $ref: "#/components/schemas/Money"
    ExpensePreview:
      description: The Shares an Expense would get, in the Split's order.
      type: object
      required: [amount, shares]
      additionalProperties: false
      properties:
        amount:
          $ref: "#/components/schemas/Money"
        shares:
          type: array
          items:
            $ref: "#/components/schemas/ShareLine"
          examples:
            - - member_id: 0190b6c4-0000-7000-8000-0000000000d1
                share: {minor: 100001, currency: INR}
...
    Money:
      description: |
        An amount in integer minor units of an ISO 4217 currency (ADR-0002).
        Never a floating-point number.
      type: object
      required: [minor, currency]
      additionalProperties: false
      properties:
        minor:
          type: integer
          format: int64
          examples: [12345]
        currency:
          $ref: "#/components/schemas/CurrencyCode"
    CurrencyCode:
      description: ISO 4217 alphabetic currency code.
      type: string
      pattern: "^[A-Z]{3}$"
      examples: ["INR"]
...
  responses:
    BadRequest:
      description: |
        The request could not be decoded (`invalid-request`), a field is
        invalid (`validation-failed`, with `errors`), or a one-time code was
        refused (`invalid-code`).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/validation-failed
            title: Validation failed
            status: 400
            errors: [{field: /password, code: too_short}]
    Unauthenticated:
      description: No access token, or it is unknown, expired or revoked.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/unauthenticated
            title: Unauthenticated
            status: 401
...
    RequestTooLarge:
      description: The request body is over 64 KB.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/request-too-large
            title: Request too large
            status: 413
...
    Forbidden:
      description: |
        The signed-in User may not do this: their email isn't verified
        (`email-not-verified`) or they aren't an Admin (`admin-required`).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/admin-required
            title: Admin required
            status: 403
    NotFound:
      description: No such resource, or one the User can't see.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/not-found
            title: Not found
            status: 404
    Conflict:
      description: |
        The request conflicts with the current state: a stale `version`
        (`version-conflict`), the Group limit (`group-limit-reached`), or a
        request with the same Idempotency-Key still running
        (`idempotency-key-in-progress`).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/version-conflict
            title: Version conflict
            status: 409
...
    InternalError:
      description: An unexpected server failure.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://splits.dev/problems/internal
            title: Internal error
            status: 500